Enhance Warning.process, bump version to 1.1.0
This handles Warning.process blocks that return :default, :backtrace,
or :raise specially. The behavior for all other return values remains
the same.  :default calls super, :backtrace calls super and prints the
backtrace to stderr, and raise raises the warning string as a
RuntimeError.

This allows Warning.process to accept a action hash instead of a
block.  Keys should be regexps or symbols recognized by Warning.ignore,
and values should be callable objects (treated the same as a
Warning.process block), or :default, :backtrace, or :raise.

This allows to easily raising exceptions for warnings or printing
backtraces for them, for specific types of warnings:

  Warning.process('', keyword_separation: :raise, taint: :backtrace)